requirements.txt
flask==3.1.0
flask-cors==5.0.0
requests==2.32.0
apscheduler==3.10.4
python-dotenv==1.0.1
gunicorn==22.0.0
